+++ addon/ng2/blueprints/component/files/__path__/__name__.model.spec.ts
@@ -1,7 +1,12 @@
 /* tslint:disable:no-unused-variable */
 
-import { addProviders, async, inject } from '@angular/core/testing';
-import {<%= classifiedModuleName %>} from './<%= fileName %>';
+import {
+  beforeEach, beforeEachProviders,
+  describe, xdescribe,
+  expect, it, xit,
+  async, inject
+} from '@angular/core/testing';
+import {<%= classifiedModuleName %>} from './<%= dasherizedModuleName %>.model';
 
 describe('<%= classifiedModuleName %>', () => {
   it('should create an instance', () => {