+++ addon/ng2/blueprints/component/files/__path__/__name__.service.spec.ts
@@ -1,16 +1,18 @@
 /* tslint:disable:no-unused-variable */
 
-import { addProviders, async, inject } from '@angular/core/testing';
+import {
+  beforeEach, beforeEachProviders,
+  describe, xdescribe,
+  expect, it, xit,
+  async, inject
+} from '@angular/core/testing';
 import { <%= classifiedModuleName %>Service } from './<%= dasherizedModuleName %>.service';
 
-describe('Service: <%= classifiedModuleName %>', () => {
-  beforeEach(() => {
-    addProviders([<%= classifiedModuleName %>Service]);
-  });
+describe('<%= classifiedModuleName %> Service', () => {
+  beforeEachProviders(() => [<%= classifiedModuleName %>Service]);
 
   it('should ...',
-    inject([<%= classifiedModuleName %>Service],
-      (service: <%= classifiedModuleName %>Service) => {
-        expect(service).toBeTruthy();
-      }));
+      inject([<%= classifiedModuleName %>Service], (service: <%= classifiedModuleName %>Service) => {
+    expect(service).toBeTruthy();
+  }));
 });